refactor: next-auth 버전 다운그레이드
auth v5에서 amplify 환경변수 인식 문제로 인한 테스트
