+++ package.json
@@ -28,7 +28,7 @@
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.390.0",
     "next": "14.2.3",
-    "next-auth": "5.0.0-beta.20",
+    "next-auth": "^4.24.7",
     "react": "^18",
     "react-dom": "^18",
     "react-slick": "^0.30.2",
