+++ src/middleware.ts
@@ -1,10 +1,10 @@
-import { NextResponse } from "next/server"
+import { NextRequest, NextResponse } from "next/server"
+import { getToken } from "next-auth/jwt"
 
-import { auth } from "./auth"
+export default async function middleware(req: NextRequest) {
+  const token = await getToken({ req })
 
-export async function middleware() {
-  const session = await auth()
-  if (!session) {
+  if (!token) {
     return NextResponse.redirect(process.env.NEXT_PUBLIC_URL as string)
   }
 }
